added login

+++ WritingPlatform/WritingPlatform.Data.Entities/Work.cs
@@ -11,7 +11,7 @@ namespace WritingPlatform.Data.Entities
 
         public DateTime PublicationTime { get; set; }
 
-        public byte Rating { get; set; }
+        public double Rating { get; set; }
 
         public int UserId { get; set; }
